Launch Gallery UI From Gallery Image Button (#35)
Created the GalleryFragment class that uses the gallery_fragment.xml layout file. Created the GalleryController that creates a dummyList tp populate the Gallery RecyclerView. Also added resouces to the res/string file and the res/color file. Added the startFragment() method in the MainActivity to start Fragments from now on.