Created the ARObjects Bottom Sheet UI (#34)
* Created the ARObjects Bottom Sheet UI

Created the Bottom Sheet for the ARObjects. Click the ARObjects Button, under the shutter to pull the ARObjects up. Also added 6 different png's for populating the Bottom Sheet. Added the ARObjects button itself, which is just drawn at runtime. Also added the RecyclerView to scroll through the ARObjects themselves. Finally, added the core library to dependencies to allow the use of BottomSheet.

* Changed ScreenOrientation

Changed the ScreenOrientation for the MainActivity.

* Changed Layout Structure

Changed the layout structure of activity_main.xml