User story 13 ar objects (#41)
* Created the ARObjects Bottom Sheet UI

Created the Bottom Sheet for the ARObjects. Click the ARObjects Button, under the shutter to pull the ARObjects up. Also added 6 different png's for populating the Bottom Sheet. Added the ARObjects button itself, which is just drawn at runtime. Also added the RecyclerView to scroll through the ARObjects themselves. Finally, added the core library to dependencies to allow the use of BottomSheet.

* Changed ScreenOrientation

Changed the ScreenOrientation for the MainActivity.

* Changed Layout Structure

Changed the layout structure of activity_main.xml

* Removed Duplicate Tags

Removed the duplicate tags in the activity_main.xml

* Changed ArObjects BottomSheet to Fragment

Changed the arobjects bottomsheet to a fragment. Fixed the bug of the bottomsheet behavior.