Sprint 2 MVP (#60)
* User story 6 load camera (#32)

* Load Camera

Loads Camera. Updates the gradle to api level 24, as well as adds dependencies to gradle and AndroidManifest.xml.

* Create Shutter Button

Created the shutter button for the camera.

* Launch Gallery UI From Gallery Image Button (#35)

Created the GalleryFragment class that uses the gallery_fragment.xml layout file. Created the GalleryController that creates a dummyList tp populate the Gallery RecyclerView. Also added resouces to the res/string file and the res/color file. Added the startFragment() method in the MainActivity to start Fragments from now on.

* Created the ARObjects Bottom Sheet UI (#34)

* Created the ARObjects Bottom Sheet UI

Created the Bottom Sheet for the ARObjects. Click the ARObjects Button, under the shutter to pull the ARObjects up. Also added 6 different png's for populating the Bottom Sheet. Added the ARObjects button itself, which is just drawn at runtime. Also added the RecyclerView to scroll through the ARObjects themselves. Finally, added the core library to dependencies to allow the use of BottomSheet.

* Changed ScreenOrientation

Changed the ScreenOrientation for the MainActivity.

* Changed Layout Structure

Changed the layout structure of activity_main.xml

* Added the Explore button to the main screen of the GARDEN. On click, it will take you to a mock Explore UI. Things to remember for next update: 1.) resize mock images in UI 2.) Resize the Explore button 3.) Change color of Explore button 4.) Decrease margins between the mock images in UI 5.) Add garden pictures to mock images in UI 6.) Look for other things

* Initial commit with the layout of the settings page with the sliders.

* Resized profile image to allow for all settings to be displayed

* Created initial fragment setup for implementation.

* Added fragment code from settings page to home screen. Need to implement fragment from home to settings page

* Added splash screen

* Added fragment functionality from homepage to settings page. This allows connectivity between the two pages.

* Revert "Merge branch 'develop' into user_story_10_setting_page"

This reverts commit 16ed1acb6c116b6bfde069217da0306593600fc7.

* Revert "Merge branch 'user_story_10_setting_page' into develop"

This reverts commit 2ff617f05e4f19e1be7863914f7e6b39cdae91de.

* Fixed layout changes to allow app to compile. Need to work on fragment usability

* Got fragment functionality working by fixing the SettingsFragment Class. SettingsController is not needed for the implementation of the fragment.

* Fixed issues from pull request comments

* splash looks less bad

* added new icons, changed icon color

* User story 1 splash screen (#36)

* Added splash screen

* splash looks less bad

* bug fixes

* bug fixes, changes to settings page and changed icons to .svg file format for crisper images

* Created the Shop Fragmnet

Created the shop fragment, allows the user to see popular porducts, recommneded products and a holiday products (Halloweeen right now). It comes with three recyclerviews that are horizontal. Also I put in four more drawable pictures.

* Created the Shop Fragmnet (#38)

Created the shop fragment, allows the user to see popular porducts, recommneded products and a holiday products (Halloweeen right now). It comes with three recyclerviews that are horizontal. Also I put in four more drawable pictures.

* Fixed Settings Page Layout Issues (#39)

* Fixed Settings Page Layout Issues

Fixed the settings page layout issues. Made the layout look cleaner, as well as fixed code layout format. Deleted unused files and import statements.

* Added Back Button To Settings

Added the back button and removed the home button from the settings fragment.

* Fixed Shopping Fragment Spacing Issues

See title

* User story 13 ar objects (#41)

* Created the ARObjects Bottom Sheet UI

Created the Bottom Sheet for the ARObjects. Click the ARObjects Button, under the shutter to pull the ARObjects up. Also added 6 different png's for populating the Bottom Sheet. Added the ARObjects button itself, which is just drawn at runtime. Also added the RecyclerView to scroll through the ARObjects themselves. Finally, added the core library to dependencies to allow the use of BottomSheet.

* Changed ScreenOrientation

Changed the ScreenOrientation for the MainActivity.

* Changed Layout Structure

Changed the layout structure of activity_main.xml

* Removed Duplicate Tags

Removed the duplicate tags in the activity_main.xml

* Changed ArObjects BottomSheet to Fragment

Changed the arobjects bottomsheet to a fragment. Fixed the bug of the bottomsheet behavior.

* Updates

Things I did:
- Added Strings for my Tabs (Home and Favorites)
- Added xml and png files for the garden logo
- Added app custom icon

Things I fixed:
- Splash Screen
- Menu Icons
- Color Scheme

* I goofed... the splash screen was drawing xml vectors for the flower which I guess you cant do. I replaced it with png.

* What I did:
Added Material as a dependency in Gradle
Added a bottom nav bar
Added a Card with mock article title
Added a Color folder with color state xml file
Added a bottom nav bar menu in menu folder

What I fixed:
Icon Sizes
Back Icon
Changed the toolbar to a Material App Bar

* Shopping page with links and pictures

* forgot some stuff

* What I did:
Added a grid system for the cards
Added images for mock data
Added Clickable Links to the data

What I fixed:
Holders and Adapters

* User story 55 settings database (#57)

* created initial event handlers for the switch components

* Added SharedPreferences to the event handlers to allow switch values to be saved in any state the app may be in.

* Applied shared preferences for all switches

* Added a Edit Text component to the settings so that users can update their usernames.

* Rearranged the UI to be a more compact so that the update username field was not in the middle of the screen

* Added a change password field that when the login page is created, change the user's password. The field has been hashed out for privacy/security reasons.

* Updated the name of the settings so that they are not generic names. This is for Task #45

* Renamed the switches and text views to what the settings were to have a consistent naming scheme.

* Task #45: Updated the component variables to private to be secure as mention in the pull request

* Add Article Covers

* Fixed errors in settings