User story 62 ar objects (#75)
* AR Objects

Added a bunch of 3d models in the glTF file format for AR Core. All models are from sketchfab.com.

* Extracted all files to allow for placement in the view

* Loading ar objects

Trying to load in ar objects into the sceneform. Currently loading in Bhudda Head object.

Also created a sampledata directory with unzipped ar objects.

* updated file structure to include name of object's artist. added a raw directory in which i am adding .sfb files after importing sceneform asset

* rename files, add raw with sceneform assets

 updated file structure to include name of object's artist. added a raw directory in which i am adding .sfb files after importing sceneform asset

* Updates to the res/raw/ folder

Added all the .sfb sceneform assets. Had to remove some that were to big to process.

* Models have preloaded but are unable to load. Need to figure out issue with the AR fragment listener events by looking at Google's example code

* AR Objects Updates

Fixed:
- Build Errors
- AR Tab Recyler View now loads in correct image thumbnails

Added:
- More fltb objects

TODO:
- Currently working on drawing renderable objects from the /raw/ folder.
- Render scene with the camera nodes
- handle user touch

* Start of selecting logic for when an AR object is selected.

* Commenting out the changes to prevent errors during compile time

* Renamed AR Assets For Simplicity

* Loading in the AR Object into the scene

- had to move all the .sfb files into the assets/ folder
- also renamed the objects for ease of access
- added a bunch of methods to load in the object

* Recyclerview ActionListener Logic

Added the a listener to the recyclerview so that we can select what ARObject we want to place down. At the moment it will only send the user to the main_activity fragment

* Allowed for selection of AR Objects based on the selection screen

Added an interface classes to allow the object to be placed into the sceneform from the fragment. We changed how the ArrayList is made so that we can pass in URI to be parsed in the main activity. Main Activity was updated for this functionality to work.

* Bug fix from merge with develop. Forgot to remove <<<Head & {

* FIxed File Location Path

Fixed the windows style of path to a linux style of path (eg "\\" to '/')

* 3D models scaling refactor

Changed the scaling factors for the 3D models. Sized down most of the models.

* Cleaned Up Code Related to ArObjects User Story

Removed unnecessary comments and repositioned code

* Removed Duplicate ARFragment Variable

* Removed Unused File and Fixed ArFragment Bug

ARFragment was not being initalized before being used, fixed that issue.

* Removed Plane Renderer

Removed the dot for the planerenderer. Makes taking pictures better.